Replace previous UsersTable implementation :0
