+++ src/entities/index.ts
@@ -1,2 +1,1 @@
-export * from './user/hooks';
-export * from './user/ui/users-table';
+export * from './users/index';
