--- src/entities/user/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './use-sorted-users.ts';
