--- src/entities/user/hooks/use-sorted-users.ts
@@ -1,38 +1,0 @@
-import { useMemo } from 'react';
-import type { User } from '~/shared';
-
-type Direction = 'asc' | 'desc';
-type SortableKeys = 'first_name' | 'last_name' | 'email' | 'department_name' | 'position_name';
-
-const accessorMap = {
-  first_name: (u: User) => u.profile?.firstName,
-  last_name: (u: User) => u.profile?.lastName,
-  email: (u: User) => u.email,
-  department_name: (u: User) => u.departmentName,
-  position_name: (u: User) => u.positionName,
-};
-
-export function useSortedUsers(users: User[], orderBy: string, direction: Direction) {
-  return useMemo(() => {
-    if (!orderBy || !accessorMap[orderBy as SortableKeys]) return users;
-
-    const getValue = accessorMap[orderBy as SortableKeys];
-    const usersCached = users.map((user) => ({
-      user,
-      value: getValue(user),
-    }));
-
-    usersCached.sort((a, b) => {
-      const aVal = a.value;
-      const bVal = b.value;
-
-      if (aVal == null || aVal === '') return 1;
-      if (bVal == null || bVal === '') return -1;
-      if (aVal === bVal) return 0;
-
-      return direction === 'asc' ? String(aVal).localeCompare(String(bVal)) : String(bVal).localeCompare(String(aVal));
-    });
-
-    return usersCached.map(({ user }) => user);
-  }, [users, orderBy, direction]);
-}
