--- src/entities/user/ui/users-table/index.ts
@@ -1,3 +1,0 @@
-export * from './users-table.ui';
-export * from './users-table-head.ui';
-export * from './user-row.ui';
