+++ src/entities/users/index.ts
@@ -1,0 +1,1 @@
+export * from './ui/index';
