+++ src/shared/ui/index.ts
@@ -1,23 +1,25 @@
-export * from './text-field.ui';
+export * from './actions-menu';
+export * from './breadcrumbs.ui';
+export * from './button-link.ui';
 export * from './button.ui';
+export * from './count.ui';
+export * from './form-errors.ui';
+export * from './highlight.ui';
+export * from './icon-button-link.ui';
+export * from './menu-item-link.ui';
+export * from './modal.ui';
+export * from './optional-label.ui';
 export * from './password-field.ui';
+export * from './search-field.ui';
+export * from './select.ui';
+export * from './skill-bar.ui';
+export * from './tab-link.ui';
+export * from './table.ui';
+export * from './tabs.ui';
+export * from './text-field.ui';
+export * from './text.ui';
 export * from './title.ui';
 export * from './user-avatar.ui';
-export * from './text.ui';
-export * from './search-field.ui';
-export * from './form-errors.ui';
-export * from './tabs.ui';
-export * from './breadcrumbs.ui';
-export * from './skill-bar.ui';
-export * from './modal.ui';
-export * from './select.ui';
-export * from './count.ui';
-export * from './button-link.ui';
-export * from './tab-link.ui';
-export * from './menu-item-link.ui';
-export * from './actions-menu';
-export * from './icon-button-link.ui';
-export * from './table.ui';
 
 export * from './page-title.ui';
